실수연산 구현

+++ src/main/java/org/example/Lv3/OperatorType.java
@@ -2,22 +2,22 @@ package org.example.Lv3;
 
 public enum OperatorType {
     add("+"){
-        public int apply(int num1, int num2) {
+        public double apply(double num1, double num2) {
             return num1 + num2;
         }
     },
     sub("-"){
-        public int apply(int num1, int num2) {
+        public double apply(double num1, double num2) {
             return num1 - num2;
         }
     },
     mul("*"){
-        public int apply(int num1, int num2) {
+        public double apply(double num1, double num2) {
             return num1 * num2;
         }
     },
     div("/"){
-        public int apply(int num1, int num2) {
+        public double apply(double num1, double num2) {
             if (num2 == 0) {
                 throw new ArithmeticException("0으로 나눌 수 없습니다!");
             } return num1 / num2;
@@ -28,7 +28,7 @@ public enum OperatorType {
     OperatorType(String symbol) {
         this.symbol = symbol;
     }
-    public abstract int apply(int num1, int num2);
+    public abstract double apply(double num1, double num2);
 
     public static OperatorType newSymbol(String symbol) {
         for (OperatorType op : OperatorType.values()) {
